Minor Hero Class and InputManager implementations
InputManager is a singleton and uses an observer pattern. It has a
delegate for Keyboard Events and AxisEvents. Has an event for Run and
Jump so far

Hero class has a class within it that is the heros stats. The Hero is
subscribed to the Run and Jump method that the InputManager has and
applies force on its RigidBody2D accordingly