Extra InputManager Keyevents and more Hero class implementations
InputManager - Attack event, Block event and Pause event.

Hero - Now is subscribed to the Attack, Block and Pause events from the
InoutManager. Hero class unsubscribes to the events in the OnDestroy
method. Hero can run left and right and when they hit their maxVelocity
they can’t get any faster. There is also a GroundCheck method that uses
on OverlapCircle to check if the heros EdgeCollider intersects with the
“Ground” layer.