Lots of updates
Archer - now shoots in an arc. Still need to implement a way of varying
the angle he shoots at and the distance the arrow travels.

Wizard - Now shoots projectiles in a straight line.

Hero - Added more comments and more functionality to not let the player
attack and block at the same time.

ColorWheel - new script to handle the behaviour of the 3D colorWheel
surrounding the scene. Just rotates randomly as of now.

Made a bunch of materials with the gradient shader I created a put one
on each inside face of the ColorWheel