Archer Wizard and Colour Implementations
Input Manager - Now has a TrackMouse delegate thats used to track where
the mouse position is on the screen. SwitchColour and SwitchShade have
also been added for the player being able to switch colours in-game

Hero - Moved generic methods out of this class and they were put into
the HelperFunctions class. A bunch of projectile shooting was added to
the Attack method for the rangedAttack heroes.

Archer - now has a shield and a bow. The bow moves by scrolling the
mouse up and down at the players feet. the bow fires an arrow at the
angle its currently at with the help from the Arc function in the
HelperFunctions class.

CameraBehaviour - used to follow the main player in the current game.
The position has a smooth damp on it so it follows the player at a not
so jittery pace.

ColourWheel - Used to rotate the 3D colourWheel that surround the
level. Stops in the center of each background using a modulus of
current rotation angle.

Colour - has a ColorContainer class inside it with an hum for each
colourType and an array of each Color shade corresponding to the Type.
A Dictionary is also used to hold all the info in the ColorContainer
class.

Projectile - small script on projectiles that just die on collision
impact.