Networking Implementation 1
This push contains Photon and everything it comes with. The first
implementation of networking is done. Players can spawn in the same
game and can see each others movement, attack and block. Players are
able to select their hero and colour in the character select screen and
spawn their hero into the same game. I need to add a way to spawn them
in a scene with the same colour wheel maybe by disabling the wheels
logic in a game thats not the hosts.

This push also introduces The NetworkManager Class - Used to Join and
Create Network rooms for Multiplayer

The MainMenu class - For button logic in the first menu of the game

CharacterSelectScreen - Used as a bridge between the game scene and the
main menu where players select their character and colour and are
brought into a game

The HeroManager class - Manages a list of all heroes and is used to
gather hero information in other classes

LevelManager - Now instantiates the hero chosen from the select screen
in the game world when the first scene loads